[CHORE] 컴포저블 코드 파일 이동
* NoticeBar 컴포저블을 GlobalComponents.kt -> StampBoardDetailComponents.kt로 이동
