+++ app/src/main/java/com/polzzak_android/presentation/component/GlobalComponents.kt
@@ -171,44 +171,3 @@ fun ButtonPreview() {
         }
     }
 }
-
-/**
- * 알림 바
- */
-@Composable
-fun NoticeBar(
-    text: String,
-    onClick: () -> Unit
-) = Row(
-    verticalAlignment = Alignment.CenterVertically,
-    modifier = Modifier
-        .fillMaxWidth()
-        .background(color = Blue100, shape = RoundedCornerShape(8.dp))
-        .border(width = 1.dp, color = Blue700.copy(alpha = 0.16f), shape = RoundedCornerShape(8.dp))
-        .padding(horizontal = 16.dp, vertical = 12.dp)
-        .clickable(
-            interactionSource = remember { MutableInteractionSource() },
-            indication = null,
-            onClick = onClick
-        )
-) {
-    Icon(
-        imageVector = Icons.Default.Notifications,
-        contentDescription = text,
-        tint = Blue400
-    )
-    Spacer(modifier = Modifier.width(8.dp))
-    Text(
-        text = text,
-        style = PolzzakTheme.typography.semiBold14,
-        color = Blue600,
-        maxLines = 1,
-        overflow = TextOverflow.Ellipsis
-    )
-}
-
-@Preview
-@Composable
-private fun NoticeBarPreview() {
-    NoticeBar(text = "도장 요청이 있어요!", onClick = {})
-}
